Add files via upload
The class Sum_Of_Arrays defines a Java program to calculate the sum of elements in a given array. It initializes an integer array arr with a set of values. It then iterates through the array using an enhanced for loop to print each element separated by commas. Another enhanced for loop is used to calculate the sum of all elements in the array, incrementally adding each element to the variable sum. Finally, it prints out the sum of the array elements along with a message indicating it. Overall, this program efficiently computes the sum of elements in the array using simple iteration techniques in Java